Update DTO ID fields from int? to long? for 64-bit support
Changed ParentId, ComponentId, and Id properties in several
ZephyrScale.RestApi.Dtos.Cloud DTO classes from int? to long?.
This enables support for larger numeric values and improves
compatibility with systems using 64-bit identifiers. No other
logic or structure was changed.

+++ ZephyrScale.RestApi/Dtos/Cloud/TestCaseCreateRequest.cs
@@ -21,7 +21,7 @@ namespace ZephyrScale.RestApi.Dtos.Cloud
         public int? EstimatedTime { get; set; }
 
         [JsonProperty("componentId")]
-        public int? ComponentId { get; set; }
+        public long? ComponentId { get; set; }
 
         [JsonProperty("priorityName")]
         public string PriorityName { get; set; }
